Added Discord e2e tests while minding differences with existing slack tests.
