+++ test/e2e/bots_tester_test.go
@@ -8,6 +8,7 @@ import (
 	"strings"
 	"testing"
 
+	"github.com/bwmarrin/discordgo"
 	"github.com/google/uuid"
 	"github.com/slack-go/slack"
 	"github.com/stretchr/testify/assert"
@@ -25,6 +26,11 @@ type slackTester struct {
 	cfg SlackConfig
 }
 
+type discordTester struct {
+	cli *discordgo.Session
+	cfg DiscordConfig
+}
+
 func newSlackTester(slackCfg SlackConfig) (*slackTester, error) {
 	slackCli := slack.New(slackCfg.TesterAppToken)
 	_, err := slackCli.AuthTest()
@@ -33,6 +39,157 @@ func newSlackTester(slackCfg SlackConfig) (*slackTester, error) {
 	}
 
 	return &slackTester{cli: slackCli, cfg: slackCfg}, nil
+}
+
+func newDiscordTester(discordCfg DiscordConfig) (*discordTester, error) {
+	discordCli, err := discordgo.New("Bot " + discordCfg.TesterAppToken)
+	if err != nil {
+		return nil, fmt.Errorf("while creating Discord session: %w", err)
+	}
+	return &discordTester{cli: discordCli, cfg: discordCfg}, nil
+}
+
+func (d *discordTester) CreateChannel(t *testing.T) (*discordgo.Channel, func(t *testing.T)) {
+	t.Helper()
+	randomID := uuid.New()
+	channelName := fmt.Sprintf("%s-%s", channelNamePrefix, randomID.String())
+
+	t.Logf("Creating channel %q...", channelName)
+	channel, err := d.cli.GuildChannelCreate(d.cfg.GuildID, channelName, discordgo.ChannelTypeGuildText)
+	require.NoError(t, err)
+
+	t.Logf("Channel %q (ID: %q) created", channelName, channel.ID)
+
+	cleanupFn := func(t *testing.T) {
+		t.Helper()
+		t.Logf("Deleting channel %q...", channel.Name)
+		// We cannot archive a channel: https://support.discord.com/hc/en-us/community/posts/360042842012-Archive-old-chat-channels
+		_, err := d.cli.ChannelDelete(channel.ID)
+		assert.NoError(t, err)
+	}
+
+	return channel, cleanupFn
+}
+
+func (d *discordTester) FindUserIDForBot(t *testing.T) string {
+	return d.FindUserID(t, d.cfg.BotName)
+}
+
+func (d *discordTester) FindUserIDForTester(t *testing.T) string {
+	return d.FindUserID(t, d.cfg.TesterName)
+}
+
+func (d *discordTester) FindUserID(t *testing.T, name string) string {
+	t.Log("Getting users...")
+	res, err := d.cli.GuildMembersSearch(d.cfg.GuildID, name, 5)
+	require.NoError(t, err)
+
+	t.Logf("Finding user ID by name %q...", name)
+	for _, m := range res {
+		if !strings.EqualFold(name, m.User.Username) {
+			continue
+		}
+		return m.User.ID
+	}
+
+	return ""
+}
+
+func (d *discordTester) PostInitialMessage(t *testing.T, channelID string) {
+	t.Helper()
+	t.Logf("Posting welcome message for channel: %s...", channelID)
+
+	var additionalContextMsg string
+	if d.cfg.AdditionalContextMessage != "" {
+		additionalContextMsg = fmt.Sprintf("%s\n", d.cfg.AdditionalContextMessage)
+	}
+	message := fmt.Sprintf("Hello!\n%s%s", additionalContextMsg, welcomeText)
+	_, err := d.cli.ChannelMessageSend(channelID, message)
+	require.NoError(t, err)
+}
+
+func (d *discordTester) InviteBotToChannel(_ *testing.T, _, _ string) {
+	// This is not required in Discord.
+	// Bots can't "join" text channels because when you join a server you're already in every text channel.
+	// See: https://stackoverflow.com/questions/60990748/making-discord-bot-join-leave-a-channel
+}
+
+func (d *discordTester) WaitForMessagePostedRecentlyEqual(userID, channelID, expectedMsg string) error {
+	return d.WaitForMessagePosted(userID, channelID, recentMessagesLimit, func(msg *discordgo.Message) bool {
+		return strings.EqualFold(msg.Content, expectedMsg)
+	})
+}
+
+func (d *discordTester) WaitForLastMessageContains(userID, channelID string, expectedMsgSubstring string) error {
+	return d.WaitForMessagePosted(userID, channelID, 1, func(msg *discordgo.Message) bool {
+		return strings.Contains(msg.Content, expectedMsgSubstring)
+	})
+}
+
+func (d *discordTester) WaitForLastMessageEqual(userID, channelID string, expectedMsg string) error {
+	return d.WaitForMessagePosted(userID, channelID, 1, func(msg *discordgo.Message) bool {
+		return msg.Content == expectedMsg
+	})
+}
+
+func (d *discordTester) PostMessageToBot(t *testing.T, userID, channelID, command string) {
+	message := fmt.Sprintf("<@%s> %s", userID, command)
+	_, err := d.cli.ChannelMessageSend(channelID, message)
+	require.NoError(t, err)
+}
+
+func (d *discordTester) WaitForMessagePosted(userID, channelID string, limitMessages int, msgAssertFn func(msg *discordgo.Message) bool) error {
+	// To always receive message content:
+	// ensure you enable the MESSAGE CONTENT INTENT for the tester bot on the developer portal.
+	// Applications ↦ Settings ↦ Bot ↦ Privileged Gateway Intents
+	// This setting has been enforced from August 31, 2022
+
+	var fetchedMessages []*discordgo.Message
+	var lastErr error
+
+	err := wait.Poll(pollInterval, d.cfg.MessageWaitTimeout, func() (done bool, err error) {
+		messages, err := d.cli.ChannelMessages(channelID, limitMessages, "", "", "")
+		if err != nil {
+			lastErr = err
+			return false, nil
+		}
+
+		fetchedMessages = messages
+		for _, msg := range messages {
+			if msg.Author.ID != userID {
+				continue
+			}
+
+			if !msgAssertFn(msg) {
+				// different message
+				continue
+			}
+
+			return true, nil
+		}
+
+		return false, nil
+	})
+	if lastErr == nil {
+		lastErr = errors.New("message assertion function returned false")
+	}
+	if err != nil {
+		if err == wait.ErrWaitTimeout {
+			return fmt.Errorf("while waiting for condition: last error: %w; fetched messages: %s", lastErr, structDumper.Sdump(fetchedMessages))
+		}
+		return err
+	}
+
+	return nil
+}
+
+func (d *discordTester) WaitForMessagesPostedOnChannels(userID string, channelIDs []string, limitMessages int, msgAssertFn func(msg *discordgo.Message) bool) error {
+	errs := multierror.New()
+	for _, channelID := range channelIDs {
+		errs = multierror.Append(errs, d.WaitForMessagePosted(userID, channelID, limitMessages, msgAssertFn))
+	}
+
+	return errs.ErrorOrNil()
 }
 
 func (s *slackTester) CreateChannel(t *testing.T) (*slack.Channel, func(t *testing.T)) {
